Remove support for JWT proofs without a key attestation. Limit the maximum supported number of proofs to 1.
