+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/domain/CredentialRequest.kt
@@ -24,10 +24,6 @@ import com.nimbusds.jose.EncryptionMethod
 import com.nimbusds.jose.JWEAlgorithm
 import com.nimbusds.jose.jwk.AsymmetricJWK
 import com.nimbusds.jose.jwk.JWK
-import eu.europa.ec.eudi.pidissuer.adapter.out.jose.resolveDidUrl
-import eu.europa.ec.eudi.pidissuer.adapter.out.util.getOrThrow
-import java.net.URI
-import java.security.cert.X509Certificate
 
 /**
  * Proof of possession.
@@ -57,45 +53,6 @@ sealed interface UnvalidatedProof {
  * inside the issued credential
  */
 sealed interface CredentialKey {
-
-    /**
-     * If the Credential shall be bound to a DID, the kid refers to a DID URL
-     * which identifies a particular key in the DID Document that the Credential shall be bound to
-     */
-    data class DIDUrl(val did: URI, val jwk: JWK) : CredentialKey {
-        init {
-            require(!jwk.isPrivate) { "jwk must not contain a private key" }
-            require(jwk is AsymmetricJWK) { "'jwk' must be asymmetric" }
-        }
-
-        companion object {
-
-            /**
-             * Resolves the provided DID url. Currently, it supports 'key' and 'jwk' methods.
-             */
-            operator fun invoke(value: String): Either<Throwable, DIDUrl> = Either.catch {
-                val url = URI.create(value)
-                val jwk = resolveDidUrl(url).getOrThrow()
-                DIDUrl(url, jwk)
-            }
-        }
-    }
-
-    @JvmInline
-    value class Jwk(val value: JWK) : CredentialKey {
-        init {
-            require(!value.isPrivate) { "jwk must not contain a private key" }
-            require(value is AsymmetricJWK) { "'jwk' must be asymmetric" }
-        }
-    }
-
-    @JvmInline
-    value class X5c(val chain: NonEmptyList<X509Certificate>) : CredentialKey {
-        val certificate: X509Certificate
-            get() = chain.head
-
-        companion object
-    }
 
     @JvmInline
     value class AttestedKeys(val keys: NonEmptyList<JWK>) : CredentialKey {
@@ -177,7 +134,7 @@ sealed interface RequestedResponseEncryption {
  */
 sealed interface CredentialRequest {
     val format: Format
-    val unvalidatedProofs: NonEmptyList<UnvalidatedProof>
+    val unvalidatedProof: UnvalidatedProof
     val credentialResponseEncryption: RequestedResponseEncryption
 }
