+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/domain/Specs.kt
@@ -23,6 +23,9 @@ internal object OpenId4VciSpec {
     val ZIP_ALGORITHMS = setOf("DEF")
 
     const val KEY_ATTESTATION_JWT_TYPE = "key-attestation+jwt"
+    const val KEY_ATTESTATION_ATTESTED_KEYS = "attested_keys"
+    const val KEY_ATTESTATION_KEY_STORAGE = "key_storage"
+    const val KEY_ATTESTATION_USER_AUTHENTICATION = "user_authentication"
 
     const val TRANSACTION_ID = "transaction_id"
     const val INTERVAL = "interval"
