+++ src/main/kotlin/eu/europa/ec/eudi/pidissuer/port/input/IssueCredential.kt
@@ -458,7 +458,7 @@ private sealed interface UnresolvedCredentialRequest {
      */
     data class ByCredentialIdentifier(
         val credentialIdentifier: CredentialIdentifier,
-        val unvalidatedProofs: NonEmptyList<UnvalidatedProof>,
+        val unvalidatedProofs: UnvalidatedProof,
         val credentialResponseEncryption: RequestedResponseEncryption,
     ) : UnresolvedCredentialRequest
 }
@@ -485,7 +485,7 @@ private interface Validations : Raise<IssueCredentialError> {
             }
         }
 
-        val proofs =
+        val proof =
             when {
                 proofs != null -> {
                     val jwtProofs = proofs.jwtProofs?.map { UnvalidatedProof.Jwt(it) }
@@ -501,12 +501,13 @@ private interface Validations : Raise<IssueCredentialError> {
 
                     val proofs = (jwtProofs.orEmpty() + diVpProofs.orEmpty() + attestations.orEmpty()).toNonEmptyListOrNull()
                     ensureNotNull(proofs) { MissingProof }
+                    ensure(proofs.size == 1) {
+                        InvalidProof("You can provide at most 1 proof")
+                    }
+                    proofs.firstOrNull()
                 }
                 else -> raise(MissingProof)
             }
-        ensure(proofs.size <= supportedBatchIssuance.maxProofsSupported) {
-            InvalidProof("You can provide at most '${supportedBatchIssuance.maxProofsSupported}' proofs")
-        }
 
         val credentialResponseEncryption = credentialResponseEncryption?.toDomain() ?: RequestedResponseEncryption.NotRequired
         credentialResponseEncryption.ensureIsSupported(supportedEncryption)
@@ -517,8 +518,8 @@ private interface Validations : Raise<IssueCredentialError> {
             supportedCredentialConfigurations.firstOrNull { credentialConfigurationId == it.id }
                 ?.let { credentialConfiguration ->
                     val credentialRequest = when (credentialConfiguration) {
-                        is MsoMdocCredentialConfiguration -> credentialConfiguration.credentialRequest(proofs, credentialResponseEncryption)
-                        is SdJwtVcCredentialConfiguration -> credentialConfiguration.credentialRequest(proofs, credentialResponseEncryption)
+                        is MsoMdocCredentialConfiguration -> credentialConfiguration.credentialRequest(proof, credentialResponseEncryption)
+                        is SdJwtVcCredentialConfiguration -> credentialConfiguration.credentialRequest(proof, credentialResponseEncryption)
                         is JwtVcJsonCredentialConfiguration -> raise(UnsupportedCredentialType(format = JWT_VS_JSON_FORMAT))
                     }
                     UnresolvedCredentialRequest.ByCredentialConfigurationId(credentialConfigurationId, credentialRequest)
@@ -527,7 +528,7 @@ private interface Validations : Raise<IssueCredentialError> {
         fun credentialRequestByCredentialIdentifier(credentialIdentifier: String): UnresolvedCredentialRequest.ByCredentialIdentifier =
             UnresolvedCredentialRequest.ByCredentialIdentifier(
                 CredentialIdentifier(credentialIdentifier),
-                proofs,
+                proof,
                 credentialResponseEncryption,
             )
